Rather than setting up a new JITM template, use sidebar-banner and do the A/B test checks within that, returning different markup depending on the test variant. This bypasses the need to change the value on the server side.
