+++ client/blocks/jitm/index.jsx
@@ -29,10 +29,6 @@ function renderTemplate( template, props ) {
 		return <AsyncLoad { ...props } require="blocks/jitm/templates/sidebar-banner" />;
 	}
 
-	if ( template === 'upsell-nudge' ) {
-		return <AsyncLoad { ...props } require="blocks/jitm/templates/upsell-nudge" />;
-	}
-
 	return <AsyncLoad { ...props } require="blocks/jitm/templates/default" />;
 }
